Splitted API logic into an external function files (not working)
